requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.0
pytest==7.4.2
pytest-xdist==3.3.1
pytest-cov==4.1.0
httpx==0.25.0
pillow==10.0.1
reportlab==4.1.0
python-dotenv==1.0.0
